requirements.txt
pandas
numpy
gradio
torch
transformers
sentence-transformers
faiss-gpu
pdfplumber
